+++ mmms-backend/src/main/java/com/sep/mmms_backend/config/JacksonConfig.java
@@ -1,0 +1,18 @@
+package com.sep.mmms_backend.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+
+@Configuration
+public class JacksonConfig {
+    //after this config, the ObjectMapper can handle 'LocalDate' data types
+    @Bean
+    public ObjectMapper objectMapper() {
+       ObjectMapper objectMapper = new ObjectMapper();
+       objectMapper.registerModule(new JavaTimeModule());
+       return objectMapper;
+    }
+}