+++ mmms-backend/src/main/java/com/sep/mmms_backend/exception_handling/GlobalExceptionHandler.java
@@ -1,0 +1,15 @@
+package com.sep.mmms_backend.exception_handling;
+
+
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<Object> handleDataIntegrityViolation(DataIntegrityViolationException e) {
+        return ResponseEntity.internalServerError().body(null);
+    }
+}