+++ mmms-backend/src/main/java/com/sep/mmms_backend/repository/MeetingRepository.java
@@ -1,0 +1,10 @@
+package com.sep.mmms_backend.repository;
+
+import com.sep.mmms_backend.entity.Meeting;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MeetingRepository extends JpaRepository<Meeting, Integer> {
+
+}