+++ mmms-backend/src/main/java/com/sep/mmms_backend/service/MeetingService.java
@@ -1,0 +1,18 @@
+package com.sep.mmms_backend.service;
+
+import com.sep.mmms_backend.entity.Meeting;
+import com.sep.mmms_backend.repository.MeetingRepository;
+import org.springframework.stereotype.Service;
+
+@Service
+public class MeetingService {
+
+    MeetingRepository meetingRepository;
+    public MeetingService(MeetingRepository meetingRepository) {
+        this.meetingRepository = meetingRepository;
+    }
+
+    public Meeting saveNewMeeting(Meeting meeting) {
+        return meetingRepository.save(meeting);
+    }
+}