+++ mmms-backend/src/test/java/com/sep/mmms_backend/controller/RegisterUserControllerTest.java
@@ -28,11 +28,9 @@ import org.springframework.test.web.servlet.MvcResult;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(controllers={RegisterUserController.class})
 @Import(SecurityConfiguration.class)